Refactor fallback utils into smaller modules
